PMX/VMD의 NaN·Inf 및 잘못된 플래그 검증 강화
VMD show 오류가 정상 처리되던 버그 수정
BinaryReader::Read를 trivially-copyable 타입으로 제한
Bullet world를 Physics 내부로 숨기고 등록·해제·스텝 계약 통일
물리 리소스 정리를 ModelPhysicsData::Reset으로 이동
작은 모델에서 빈 스키닝 작업이 생성되지 않도록 최적화
그룹 모프 순환 검사를 재귀 없는 단일 DFS로 변경
도달 불가능한 LoadFaces 실패 분기 제거
프레임·문자열 크기 변환 경고 정리
손상 PMX/VMD와 스키닝 범위 회귀 테스트 추가

+++ src/Core/Animation/Camera/CameraAnimation.cpp
@@ -25,8 +25,8 @@ namespace Chrivent {
 			camera.fov = prev.fov;
 			return camera;
 		}
-		const float prevFrame = prev.frame;
-		const float nextFrame = frame;
+		const float prevFrame = static_cast<float>(prev.frame);
+		const float nextFrame = static_cast<float>(frame);
 		const float normalizedTime = (t - prevFrame) / (nextFrame - prevFrame);
 		const float ixY = ixBezier.Evaluate(normalizedTime);
 		const float iyY = iyBezier.Evaluate(normalizedTime);
